ppt_remote/ppt_remote_python/requirements.txt
pyautogui==0.9.54
Pillow==10.0.0
keyboard==0.13.5 